Add options --wrap, --no-wrap and --width=NUMBER.
Use these to determine whether long lines are wrapped and at which width.
Default width is 79.
By default we do NOT wrap, because we have never wrapped before.
This may change in the future, so if you *really* want to be sure to not wrap
when using a future i18ndude version, you can add --no-wrap now.

Note that we no longer read the PO_MAX_WIDTH environment variable,
which we did in an earlier version of this wrapping feature.
See https://github.com/collective/i18ndude/issues/3